requirements.txt
pandas==0.23.0
numpy==1.14.3
Keras==2.1.6
scikit_image==0.13.0
tqdm==4.23.4
matplotlib==2.2.2
scipy==1.1.0
tensorflow-gpu==1.1.0
opencv_python==3.4.1.15
Pillow==5.2.0
skimage==0.0
scikit_learn==0.19.1
PyDrive==1.3.1
imgaug
